[CHORE]: 폴더명 Snake Case로 변경
