+++ modules/domain/src/main/java/com/ott/domain/ingest_job/domain/IngestJob.java
@@ -1,10 +1,12 @@
-package com.ott.domain.watchhistory.domain;
+package com.ott.domain.ingest_job.domain;
 
 import com.ott.domain.common.BaseEntity;
 import com.ott.domain.contents.domain.Contents;
-import com.ott.domain.member.domain.Member;
+import com.ott.domain.short_form.domain.ShortForm;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -18,28 +20,27 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 @Builder
 @Getter
-@Table(name = "watch_history")
-public class WatchHistory extends BaseEntity {
+@Table(name = "ingest_job")
+public class IngestJob extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
+    @JoinColumn(name = "short_form_id")
+    private ShortForm shortForm;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "contents_id", nullable = false)
+    @JoinColumn(name = "contents_id")
     private Contents contents;
 
-    @Column(name = "last_watched_at")
-    private LocalDateTime lastWatchedAt;
+    @Enumerated(EnumType.STRING)
+    @Column(name = "ingest_status", nullable = false)
+    private IngestStatus ingestStatus;
 }
