+++ modules/domain/src/main/java/com/ott/domain/watch_history/domain/WatchHistory.java
@@ -1,8 +1,9 @@
-package com.ott.domain.seriestag.domain;
+package com.ott.domain.watch_history.domain;
 
 import com.ott.domain.common.BaseEntity;
-import com.ott.domain.series.domain.Series;
-import com.ott.domain.tag.domain.Tag;
+import com.ott.domain.contents.domain.Contents;
+import com.ott.domain.member.domain.Member;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -17,23 +18,28 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 @Builder
 @Getter
-@Table(name = "series_tag")
-public class SeriesTag extends BaseEntity {
+@Table(name = "watch_history")
+public class WatchHistory extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "tag_id", nullable = false)
-    private Tag tag;
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "series_id", nullable = false)
-    private Series series;
+    @JoinColumn(name = "contents_id", nullable = false)
+    private Contents contents;
+
+    @Column(name = "last_watched_at")
+    private LocalDateTime lastWatchedAt;
 }
